// src/app/core/api/v1/model/postsSearchRequest.ts
/**
 * OpenAPI Document on Azure Functions Demo by Adam Cox (Vorba Corp)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: adam.cox@vorba.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PostsSearchRequest { 
    blogId?: number;
    q?: string;
    pageNumber?: number;
    pageSize?: number;
}